LayoutTests/media/video-dom-end.html

<video autoplay src=content/test.mp4 controls></video>
<script src=video-test.js></script>
<script>
    video.end = 0.5;
    testExpected("video.end", 0.5);
    testExpected("video.getAttribute('end')", "0.5s");

    waitForEvent('ended', function () { 
        testExpected("video.currentTime", 0.5);
        endTest();
    } );
</script>
