LayoutTests/media/video-dom-end-expected.txt

EXPECTED (video.end == '0.5') OK
EXPECTED (video.getAttribute('end') == '0.5s') OK
EVENT(ended)
EXPECTED (video.currentTime == '0.5') OK
END OF TEST
